origraph-registry-demo/requirements.txt
fastapi
uvicorn[standard]
anthropic
python-dotenv
mangum
boto3
eth-account>=0.13
eth-keys>=0.6
-e ../invisible-text-watermark
pytest
